--- composeApp/src/commonMain/kotlin/com/github/familyvault/models/chat/ChatMessageImageMetadata.kt
@@ -1,6 +1,0 @@
-package com.github.familyvault.models.chat
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class ChatMessageImageMetadata(val storeFileId: String, val height: Int, val width: Int)
