+++ composeApp/src/commonMain/kotlin/com/github/familyvault/models/enums/StoreType.kt
@@ -2,5 +2,5 @@ package com.github.familyvault.models.enums
 
 enum class StoreType {
     CHAT_FILES,
-    FILE_CABINET
+    FILE_CABINET_IMAGES
 }
