+++ composeApp/src/commonMain/kotlin/com/github/familyvault/services/FamilyMemberAdditionService.kt
@@ -20,7 +20,7 @@ class FamilyMemberAdditionService(
             )
         )
 
-        fileCabinetService.updateFamilyGroupFileCabinet()
+        fileCabinetService.restoreFileCabinetMembership()
         taskService.restoreTaskListsMembership()
     }
