+++ composeApp/src/commonMain/kotlin/com/github/familyvault/ui/screens/main/filesCabinet/PhotosTabContent.kt
@@ -1,15 +1,10 @@
 package com.github.familyvault.ui.screens.main.filesCabinet
 
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Card
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.foundation.lazy.grid.GridCells
 import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
-import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -17,16 +12,15 @@ import androidx.compose.runtime.produceState
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.ImageBitmap
-import androidx.compose.ui.layout.ContentScale
 import com.github.familyvault.services.IFileCabinetService
 import com.github.familyvault.services.IImagePickerService
 import com.github.familyvault.ui.components.FullScreenImage
 import com.github.familyvault.ui.components.LoaderWithText
+import com.github.familyvault.ui.components.filesCabinet.LoadingCard
+import com.github.familyvault.ui.components.filesCabinet.PhotoCard
 import com.github.familyvault.ui.theme.AdditionalTheme
 import familyvault.composeapp.generated.resources.Res
-import familyvault.composeapp.generated.resources.file_cabinet_photo_card
 import familyvault.composeapp.generated.resources.loading
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.IO
@@ -94,47 +88,6 @@ fun PhotosTabContent() {
         FullScreenImage(
             imageBitmap = fullScreenImage!!,
             onDismiss = { fullScreenImage = null }
-        )
-    }
-}
-
-@Composable
-fun PhotoCard(
-    imageBitmap: ImageBitmap,
-    onClick: () -> Unit
-) {
-    Card(
-        modifier = Modifier
-            .padding(AdditionalTheme.spacings.small)
-            .fillMaxWidth()
-            .clickable(onClick = onClick),
-        shape = RoundedCornerShape(AdditionalTheme.roundness.normalPercent),
-    ) {
-        Image(
-            bitmap = imageBitmap,
-            contentDescription = stringResource(Res.string.file_cabinet_photo_card),
-            modifier = Modifier
-                .fillMaxWidth()
-                .height(AdditionalTheme.sizing.fileCabinetNormal)
-                .padding(AdditionalTheme.spacings.small)
-                .clip(RoundedCornerShape(AdditionalTheme.roundness.normalPercent)),
-            contentScale = ContentScale.Crop
-        )
-    }
-}
-
-@Composable
-fun LoadingCard() {
-    Card(
-        modifier = Modifier
-            .padding(AdditionalTheme.spacings.medium)
-            .fillMaxWidth()
-            .height(AdditionalTheme.sizing.fileCabinetNormal),
-        shape = RoundedCornerShape(AdditionalTheme.roundness.normalPercent)
-    ) {
-        Box(
-            modifier = Modifier
-                .fillMaxSize(),
         )
     }
 }
